// circabc-frontend/src/app/core/generated/circabc/model/eventDefinition.ts
/**
 * CIRCABC REST API
 * This is the first version of the CIRCABC REST API used by the new User Interface 
 *
 * The version of the OpenAPI document: 4.2.0
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DateInfo } from './dateInfo';
import { RepeatsInfo } from './repeatsInfo';
import { AttendantsInfo } from './attendantsInfo';


/**
 * definition of an event or meeting. Used to create and retrieve information about it
 */
export interface EventDefinition { 
    appointmentTypeEvent: boolean;
    id?: string;
    igId?: string;
    title: string;
    dateInfo: DateInfo;
    language: string;
    timezone: string;
    eventAbstract?: string;
    invitationMessage?: string;
    location?: string;
    repeatsInfo: RepeatsInfo;
    attendantsInfo: AttendantsInfo;
    enableNotification?: boolean;
    contactName: string;
    contactPhone: string;
    contactEmail: string;
    contactUrl?: string;
    meetingPublicAvailability: boolean;
    meetingOrganisation?: string;
    meetingAgenda?: string;
    meetingType: string;
    meetingLibrarySection?: string;
    eventType: EventDefinition.EventTypeEnum;
    eventPriority: EventDefinition.EventPriorityEnum;
}
export namespace EventDefinition {
    export type EventTypeEnum = 'Task' | 'Appointment' | 'Other';
    export const EventTypeEnum = {
        Task: 'Task' as EventTypeEnum,
        Appointment: 'Appointment' as EventTypeEnum,
        Other: 'Other' as EventTypeEnum
    };
    export type EventPriorityEnum = 'Low' | 'Medium' | 'High' | 'Urgent';
    export const EventPriorityEnum = {
        Low: 'Low' as EventPriorityEnum,
        Medium: 'Medium' as EventPriorityEnum,
        High: 'High' as EventPriorityEnum,
        Urgent: 'Urgent' as EventPriorityEnum
    };
}