// circabc-frontend/src/app/core/generated/circabc/model/interestGroup.ts
/**
 * CIRCABC REST API
 * This is the first version of the CIRCABC REST API used by the new User Interface 
 *
 * The version of the OpenAPI document: 4.2.0
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Representation of an Interest group in CIRCABC in a JSON object
 */
export interface InterestGroup { 
    readonly id?: string;
    name: string;
    /**
     * The object that is used to compile all the translations of a node into a JSON object. It is basically composed of a map with a key languaguage code and its value 
     */
    title?: { [key: string]: string; };
    /**
     * The object that is used to compile all the translations of a node into a JSON object. It is basically composed of a map with a key languaguage code and its value 
     */
    description?: { [key: string]: string; };
    /**
     * The object that is used to compile all the translations of a node into a JSON object. It is basically composed of a map with a key languaguage code and its value 
     */
    contact?: { [key: string]: string; };
    libraryId?: string;
    informationId?: string;
    newsgroupId?: string;
    eventId?: string;
    isPublic?: boolean;
    isRegistered?: boolean;
    allowApply?: boolean;
    /**
     * Representation of the permissions given to the user making the call. To know what are the permissions for each services inside one group. 
     */
    permissions: { [key: string]: string; };
    logoUrl?: string;
}